Move classes back to original place
