+++ src/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/pipeline/LegacyLauncher.java
@@ -23,8 +23,8 @@ import fr.inria.spirals.repairnator.serializer.*;
 import fr.inria.spirals.repairnator.serializer.engines.SerializerEngine;
 import fr.inria.spirals.repairnator.states.LauncherMode;
 import fr.inria.spirals.repairnator.states.ScannedBuildStatus;
-import fr.inria.spirals.repairnator.utils.LauncherUtils;
-import fr.inria.spirals.repairnator.utils.TravisLauncherUtils;
+import fr.inria.spirals.repairnator.LauncherUtils;
+import fr.inria.spirals.repairnator.TravisLauncherUtils;
 import fr.inria.spirals.repairnator.utils.Utils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
