+++ src/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/pipeline/travis/TravisInitConfig.java
@@ -6,8 +6,8 @@ import fr.inria.spirals.repairnator.config.RepairnatorConfig;
 import fr.inria.spirals.repairnator.pipeline.IInitConfig;
 import fr.inria.spirals.repairnator.pipeline.LauncherHelpers;
 import fr.inria.spirals.repairnator.states.LauncherMode;
-import fr.inria.spirals.repairnator.utils.LauncherUtils;
-import fr.inria.spirals.repairnator.utils.TravisLauncherUtils;
+import fr.inria.spirals.repairnator.LauncherUtils;
+import fr.inria.spirals.repairnator.TravisLauncherUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
